add every room (hopefully fixed http error)

+++ app/src/main/java/at/fhooe/mc/android/fhroomfinder/ReceiveFileTask.java
@@ -9,58 +9,67 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.lang.ref.WeakReference;
+import java.net.HttpURLConnection;
 import java.net.URL;
-import java.net.URLConnection;
 
 public class ReceiveFileTask extends AsyncTask<String, Void, Void> {
 
-    private Context mContext; // TODO leak?
+    private WeakReference<Context> mContext; // TODO leak?
     private TimetableFragment ttFragment;
 
-    ReceiveFileTask(Context c, TimetableFragment timetableFragment) {
-        mContext = c;
-        ttFragment = timetableFragment;
+    ReceiveFileTask(Context _c, TimetableFragment _timetableFragment) {
+        mContext = new WeakReference<>(_c);
+        ttFragment = _timetableFragment;
     }
 
     @Override
     protected Void doInBackground(String... _url) {
-        File f = new File(mContext.getFilesDir(), "ical.ics");
-        f.delete();
+        Context context = mContext.get();
+        if (context != null) {
+            File f = new File(context.getFilesDir(), "ical.ics");
+            f.delete();
 
-        int count;
-        try {
-            URL url = new URL(_url[0]);
-            URLConnection conection = url.openConnection();
-            conection.connect();
+            int count;
+            try {
+                URL url = new URL(_url[0]);
+                HttpURLConnection connection = (HttpURLConnection) url.openConnection();
+                connection.connect();
 
-            Log.i("FHRoomFinder", "starting to download"); // TODO remove
+                // 20 sec timeout
+                connection.setConnectTimeout(20000);
+                connection.setReadTimeout(20000);
 
-            // download the file
-            InputStream input = new BufferedInputStream(url.openStream(),
-                    8192);
+                Log.i("FHRoomFinder", "starting to download"); // TODO remove
 
-            File file = new File(mContext.getFilesDir(), "ical.ics");
-            OutputStream output = new FileOutputStream(file);
+                // download the file
+                InputStream input = new BufferedInputStream(url.openStream(),
+                        8192);
 
-            byte data[] = new byte[1024];
+                File file = new File(context.getFilesDir(), "ical.ics");
+                OutputStream output = new FileOutputStream(file);
 
-            while ((count = input.read(data)) != -1) {
-                // writing data to file
-                output.write(data, 0, count);
-            }
+                byte data[] = new byte[1024];
+                while ((count = input.read(data)) != -1) {
+                    // writing data to file
+                    output.write(data, 0, count);
+                }
 
-            output.flush();
-            output.close();
-            input.close();
+                output.flush();
+                output.close();
+                input.close();
 
-            Log.i("FHRoomFinder", "file downloaded!"); // TODO remove
-            ttFragment.setInvalidURL(false);
-            ttFragment.readCal();
+                Log.i("FHRoomFinder", "file downloaded!"); // TODO remove
+                ttFragment.setInvalidURL(false);
+                ttFragment.readCal();
 
-        } catch (Exception e) {
-            Log.e("FHRoomFinder", e.getMessage());
-            if (e.getMessage().contains("no protocol")) {
-                ttFragment.setInvalidURL(true);
+                connection.disconnect();
+
+            } catch (Exception e) {
+                Log.e("FHRoomFinder", e.getMessage());
+                if (e.getMessage().contains("no protocol")) {
+                    ttFragment.setInvalidURL(true);
+                }
             }
         }
         return null;
@@ -68,6 +77,7 @@ public class ReceiveFileTask extends AsyncTask<String, Void, Void> {
 
     @Override
     protected void onPostExecute(Void aVoid) {
-        ttFragment.updateUI();
+        if (!ttFragment.isDetached())
+            ttFragment.updateUI();
     }
 }
